[create] Move fs commands to write to seeder file out of the server file and into its own file
