--- seeder.js
@@ -1,192 +1,0 @@
-[{
-    title: "The Republic",
-    author: "Plato",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The Odyssey: Rendered into English prose for the use of those who cannot read the original",
-    author: "Homer",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The Iliad",
-    author: "Homer",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The Iliad",
-    author: "Homer",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The History of the Peloponnesian War",
-    author: "Thucydides",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The Republic of Plato",
-    author: "Plato",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The Poetics of Aristotle",
-    author: "Aristotle",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Medea of Euripides",
-    author: "Euripides",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Oedipus King of Thebes: Translated into English Rhyming Verse with Explanatory Notes",
-    author: "Sophocles",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The Odyssey",
-    author: "Homer",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Plays of Sophocles: Oedipus the King; Oedipus at Colonus; Antigone",
-    author: "Sophocles",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The History of Herodotus — Volume 1",
-    author: "Herodotus",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The Aeneid",
-    author: "Virgil",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Lysistrata",
-    author: "Aristophanes",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Apology",
-    author: "Plato",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The Iliad of Homer: Translated into English Blank Verse by William Cowper",
-    author: "Homer",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Gorgias",
-    author: "Plato",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The Ethics of Aristotle",
-    author: "Aristotle",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Anabasis",
-    author: "Xenophon",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Symposium",
-    author: "Plato",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Phaedo",
-    author: "Plato",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Apology, Crito, and Phaedo of Socrates",
-    author: "Plato",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Cicero's Tusculan Disputations: Also, Treatises On The Nature Of The Gods, And On The Commonwealth",
-    author: "Cicero, Marcus Tullius",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Euthyphro",
-    author: "Plato",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Politics: A Treatise on Government",
-    author: "Aristotle",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The Bacchae of Euripides",
-    author: "Euripides",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Cyropaedia: The Education of Cyrus",
-    author: "Xenophon",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Laws",
-    author: "Plato",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "On the Nature of Things",
-    author: "Lucretius Carus, Titus",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Timaeus",
-    author: "Plato",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "Theaetetus",
-    author: "Plato",
-    createdAt: new Date(),
-    updatedAt: new Date()
-},
-{
-    title: "The Trojan Women of Euripides",
-    author: "Euripides",
-    createdAt: new Date(),
-    updatedAt: new Date()
-}]
